Upgrade to lazy-microstates api.
The next version of microstates will probably not allow access of the
`state` property from non-leaf nodes, so usages like
`dosage.state.dosage` will be broken and need to be re-written as
`dosage.dosage.state`

See https://github.com/microstates/microstates.js/pull/222 for details.

The good news is that this works with the current version of
microstates, so upgrading is not required. But it is forward
compatible with the new syntax.